Addressed comments and cleaned up
