+++ holoviews/core/data/__init__.py
@@ -8,12 +8,11 @@ except ImportError:
 import numpy as np
 import param
 
+from .. import util
 from ..dimension import redim, Dimension, process_dimensions
 from ..element import Element
 from ..ndmapping import OrderedDict
 from ..spaces import HoloMap, DynamicMap
-from ..util import (basestring, dimension_range as d_range, get_param_values,
-                    isfinite, process_ellipses, unique_iterator, wrap_tuple)
 from .interface import Interface, iloc, ndloc
 from .array import ArrayInterface
 from .dictionary import DictInterface
@@ -208,7 +207,7 @@ class Dataset(Element):
 
     def __init__(self, data, kdims=None, vdims=None, **kwargs):
         if isinstance(data, Element):
-            pvals = get_param_values(data)
+            pvals = util.get_param_values(data)
             kwargs.update([(l, pvals[l]) for l in ['group', 'label']
                            if l in pvals and l not in kwargs])
         kwargs.update(process_dimensions(kdims, vdims))
@@ -279,7 +278,7 @@ class Dataset(Element):
 
         if dim is None or (not data_range and not dimension_range):
             return (None, None)
-        elif all(isfinite(v) for v in dim.range) and dimension_range:
+        elif all(util.isfinite(v) for v in dim.range) and dimension_range:
             return dim.range
         elif dim in self.dimensions() and data_range and len(self):
             lower, upper = self.interface.range(self, dim)
@@ -287,7 +286,7 @@ class Dataset(Element):
             lower, upper = (np.NaN, np.NaN)
         if not dimension_range:
             return lower, upper
-        return d_range(lower, upper, dim.range, dim.soft_range)
+        return util.dimension_range(lower, upper, dim.range, dim.soft_range)
 
 
     def add_dimension(self, dimension, dim_pos, dim_val, vdim=False, **kwargs):
@@ -297,7 +296,7 @@ class Dataset(Element):
         dimensions and a key value scalar or sequence of the same length
         as the existing keys.
         """
-        if isinstance(dimension, (basestring, tuple)):
+        if isinstance(dimension, (util.basestring, tuple)):
             dimension = Dimension(dimension)
 
         if dimension.name in self.kdims:
@@ -396,7 +395,7 @@ class Dataset(Element):
            (4) A boolean array index matching the length of the Dataset
                object.
         """
-        slices = process_ellipses(self, slices, vdim_selection=True)
+        slices = util.process_ellipses(self, slices, vdim_selection=True)
         if isinstance(slices, np.ndarray) and slices.dtype.kind == 'b':
             if not len(slices) == len(self):
                 raise IndexError("Boolean index must match length of sliced object")
@@ -464,11 +463,11 @@ class Dataset(Element):
                 reindexed = selection.clone(new_type=Dataset).reindex(kdims)
                 selection = tuple(reindexed.columns(kdims+self.vdims).values())
 
-            datatype = list(unique_iterator(self.datatype+['dataframe', 'dict']))
+            datatype = list(util.unique_iterator(self.datatype+['dataframe', 'dict']))
             return self.clone(selection, kdims=kdims, new_type=new_type,
                               datatype=datatype)
 
-        lens = set(len(wrap_tuple(s)) for s in samples)
+        lens = set(len(util.wrap_tuple(s)) for s in samples)
         if len(lens) > 1:
             raise IndexError('Sample coordinates must all be of the same length.')
 
@@ -477,7 +476,7 @@ class Dataset(Element):
                 samples = self.closest(samples)
             except NotImplementedError:
                 pass
-        samples = [wrap_tuple(s) for s in samples]
+        samples = [util.wrap_tuple(s) for s in samples]
         return self.clone(self.interface.sample(self, samples), new_type=Table)
 
 
@@ -572,7 +571,7 @@ class Dataset(Element):
             group_dims = [kd for kd in self.kdims if kd not in dimensions]
             kdims = [self.get_dimension(d) for d in kwargs.pop('kdims', group_dims)]
             drop_dim = len(group_dims) != len(kdims)
-            group_kwargs = dict(get_param_values(self), kdims=kdims)
+            group_kwargs = dict(util.get_param_values(self), kdims=kdims)
             group_kwargs.update(kwargs)
             def load_subset(*args):
                 constraint = dict(zip(dim_names, args))
@@ -668,7 +667,7 @@ class Dataset(Element):
         """
         if 'datatype' not in overrides:
             datatypes = [self.interface.datatype] + self.datatype
-            overrides['datatype'] = list(unique_iterator(datatypes))
+            overrides['datatype'] = list(util.unique_iterator(datatypes))
         return super(Dataset, self).clone(data, shared_data, new_type, *args, **overrides)
