+++ holoviews/core/util.py
@@ -1269,10 +1269,9 @@ def is_dataframe(data):
     """
     Checks whether the supplied data is of DataFrame type.
     """
+    dd = None
     if 'dask' in sys.modules:
         import dask.dataframe as dd
-    else:
-        dd = None
     return((pd is not None and isinstance(data, pd.DataFrame)) or
           (dd is not None and isinstance(data, dd.DataFrame)))
 
@@ -1281,10 +1280,9 @@ def is_series(data):
     """
     Checks whether the supplied data is of Series type.
     """
+    dd = None
     if 'dask' in sys.modules:
         import dask.dataframe as dd
-    else:
-        dd = None
     return((pd is not None and isinstance(data, pd.Series)) or
           (dd is not None and isinstance(data, dd.Series)))
